requirements.txt
anyio==3.2.1
flatbuffers==2.0
idna==3.2
imageio==2.9.0
numpy==1.21.0
onnx==1.9.0
onnxruntime==1.8.1
opencv-python==4.5.3.56
Pillow==8.3.1
protobuf==3.17.3
six==1.16.0
sniffio==1.2.0
starlette==0.15.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
